requirements.txt
google-cloud-storage
google-cloud-bigquery
google-auth
pandas
numpy
dbt-bigquery
apache-airflow
apache-airflow-providers-google
pyyaml
dotenv
